'养修预约-结算单-前端完成-解决项目模块循环-统一合并到预约模块'

+++ wolfcode-car-appointment/src/main/java/com/wolfcode/car/appointment/domain/BusAppointment.java
@@ -43,7 +43,7 @@ public class BusAppointment implements Serializable {
 
     /** 服务类型【维修0/保养1】 */
     @Excel(name = "服务类型【维修0/保养1】")
-    private Integer serviceType;
+    private Long serviceType;
 
     /** 创建时间 */
     @JsonFormat(pattern = "yyyy-MM-dd")
